requirements.txt
beautifulsoup4==4.12.3
openai==1.30.5
Requests==2.32.3
streamlit==1.35.0
youtube_transcript_api==0.6.2
anthropic==0.10.1
google-generativeai